# Support mmcv bricks in flops compuation