Fix https://github.com/open-mmlab/mmcv/issues/675

Running the script in that issue

before:

```bash
Sequential(
  (0): Conv2d(30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
Sequential(
  (0): Conv2d(30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
Sequential(
  0.008 M, 50.000% Params, 0.531 GFLOPs, 100.000% FLOPs, 
  (0): Conv2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 100.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(0.0 M, 0.000% Params, 0.0 GFLOPs, 0.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
Sequential(
  0.016 M, 100.000% Params, 1.062 GFLOPs, 100.000% FLOPs, 
  (0): Conv2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 50.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 50.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
```

after:

```bash
Sequential(
  (0): Conv2d(30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
Sequential(
  (0): Conv2d(30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
Sequential(
  0.016 M, 100.000% Params, 1.062 GFLOPs, 100.000% FLOPs, 
  (0): Conv2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 50.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 50.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
Sequential(
  0.016 M, 100.000% Params, 1.062 GFLOPs, 100.000% FLOPs, 
  (0): Conv2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 50.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(1, 1), padding=(1, 1), bias=False)
  (1): ConvTranspose2d(0.008 M, 50.000% Params, 0.531 GFLOPs, 50.000% FLOPs, 30, 30, kernel_size=(3, 3), stride=(2, 2), output_padding=(1, 1), bias=False)
)
```